[refactor]: PostContext 생성
