+++ src/pages/PostsManagerPage.tsx
@@ -17,15 +17,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "../shared/ui";
-import {
-  deleteExistingPost,
-  getPosts,
-  getPostsByTag,
-  getSearchPosts,
-  getTags,
-  postNewPost,
-  putExistingPost,
-} from "../entities/post/api";
+
 import {
   deleteExistingComment,
   getComments,
@@ -37,29 +29,10 @@ import { UserModal } from "../feature/user/ui";
 import { useUser } from "../feature/user/model";
 import { AddPostDialog, DetailPostDialog, EditPostDialog, PostTable } from "../feature/post/ui";
 import { AddCommentDialog, EditCommentDialog } from "../feature/comment/ui";
+import { usePostContext } from "../feature/post/model/PostContext.tsx";
 
 // post게시물, comment, user
 const PostsManager = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const queryParams = new URLSearchParams(location.search);
-
-  // 상태 관리
-  const [posts, setPosts] = useState<Post[]>([]);
-  const [total, setTotal] = useState<number>(0);
-  const [skip, setSkip] = useState<number>(parseInt(queryParams.get("skip") || "0"));
-  const [limit, setLimit] = useState<number>(parseInt(queryParams.get("limit") || "10"));
-  const [searchQuery, setSearchQuery] = useState<string>(queryParams.get("search") || "");
-  const [selectedPost, setSelectedPost] = useState<Post>();
-  const [sortBy, setSortBy] = useState<string>(queryParams.get("sortBy") || "");
-  const [sortOrder, setSortOrder] = useState<string>(queryParams.get("sortOrder") || "asc");
-  const [showAddDialog, setShowAddDialog] = useState<boolean>(false);
-  const [showEditDialog, setShowEditDialog] = useState<boolean>(false);
-  const [newPost, setNewPost] = useState<NewPost>({ title: "", body: "", userId: 1 });
-  const [loading, setLoading] = useState<boolean>(false);
-  const [tags, setTags] = useState<Tag[]>([]);
-  const [selectedTag, setSelectedTag] = useState<string>(queryParams.get("tag") || "");
-
   //comment
   const [comments, setComments] = useState<Record<number, Comments[]>>([]);
   const [selectedComment, setSelectedComment] = useState<Comments>();
@@ -67,127 +40,44 @@ const PostsManager = () => {
 
   const [showAddCommentDialog, setShowAddCommentDialog] = useState<boolean>(false);
   const [showEditCommentDialog, setShowEditCommentDialog] = useState<boolean>(false);
-  const [showPostDetailDialog, setShowPostDetailDialog] = useState<boolean>(false);
+
+  const {
+    posts,
+    total,
+    loading,
+    tags,
+    selectedTag,
+    selectedPost,
+    newPost,
+    showAddDialog,
+    showEditDialog,
+    skip,
+    limit,
+    searchQuery,
+    sortBy,
+    sortOrder,
+    updateURL,
+    searchPosts,
+    fetchPostsByTag,
+    addPost,
+    updatePost,
+    deletePost,
+    setSkip,
+    setLimit,
+    setSearchQuery,
+    setSortBy,
+    setSortOrder,
+    setSelectedTag,
+    setSelectedPost,
+    setNewPost,
+    setShowAddDialog,
+    setShowEditDialog,
+  } = usePostContext();
 
   //user
   const { showUserModal, selectedUser, setShowUserModal, openUserModal } = useUser();
 
-  // URL 업데이트 함수
-  const updateURL = () => {
-    const params = new URLSearchParams();
-    if (skip) params.set("skip", skip.toString());
-    if (limit) params.set("limit", limit.toString());
-    if (searchQuery) params.set("search", searchQuery);
-    if (sortBy) params.set("sortBy", sortBy);
-    if (sortOrder) params.set("sortOrder", sortOrder);
-    if (selectedTag) params.set("tag", selectedTag);
-    navigate(`?${params.toString()}`);
-  };
-
-  // 게시물 가져오기
-  const fetchPosts = async () => {
-    setLoading(true);
-
-    const response = await getPosts(limit, skip);
-
-    if (response) {
-      const { postsData, usersData } = response;
-
-      const postsWithUsers = postsData.posts.map((post: Post) => ({
-        ...post,
-        author: usersData.users.find((user: User) => user.id === post.userId),
-      }));
-
-      setPosts(postsWithUsers);
-      setTotal(postsData.total);
-    }
-
-    setLoading(false);
-  };
-
-  // 태그 가져오기
-  const fetchTags = async () => {
-    const data = await getTags();
-
-    if (data) {
-      setTags(data);
-    }
-  };
-
-  // // 게시물 검색
-  const searchPosts = async () => {
-    if (!searchQuery) {
-      await fetchPosts();
-      return;
-    }
-
-    setLoading(true);
-
-    const data = await getSearchPosts(searchQuery);
-
-    if (data) {
-      setPosts(data.posts);
-      setTotal(data.total);
-
-      setLoading(false);
-    }
-  };
-
-  // 태그별 게시물 가져오기
-  const fetchPostsByTag = async (tag: string) => {
-    if (!tag || tag === "all") {
-      await fetchPosts();
-      return;
-    }
-
-    setLoading(true);
-
-    const data = await getPostsByTag(tag);
-
-    if (data) {
-      const { postsData, usersData } = data;
-
-      const postsWithUsers = postsData.posts.map((post: Post) => ({
-        ...post,
-        author: usersData.users.find((user: User) => user.id === post.userId),
-      }));
-
-      setPosts(postsWithUsers);
-      setTotal(postsData.total);
-    }
-
-    setLoading(false);
-  };
-
-  // 게시물 추가
-  const addPost = async () => {
-    const data = await postNewPost(newPost);
-
-    if (data) {
-      setPosts([data, ...posts]);
-      setShowAddDialog(false);
-      setNewPost({ title: "", body: "", userId: 1 });
-    }
-  };
-
-  // 게시물 업데이트
-  const updatePost = async () => {
-    if (selectedPost) {
-      const data = await putExistingPost(selectedPost);
-
-      if (data) {
-        setPosts(posts.map((post) => (post.id === data.id ? data : post)));
-        setShowEditDialog(false);
-      }
-    }
-  };
-
-  // 게시물 삭제
-  const deletePost = async (id: number) => {
-    await deleteExistingPost(id);
-
-    setPosts(posts.filter((post) => post?.id !== id));
-  };
+  const [showPostDetailDialog, setShowPostDetailDialog] = useState<boolean>(false);
 
   // 댓글 가져오기
   const fetchComments = async (postId: number) => {
@@ -259,33 +149,10 @@ const PostsManager = () => {
 
   // 게시물 상세 보기
   const openPostDetail = async (post: Post) => {
-    setSelectedPost(post);
+    setSelectedPost({ ...post });
     await fetchComments(post.id);
     setShowPostDetailDialog(true);
   };
-
-  useEffect(() => {
-    fetchTags();
-  }, []);
-
-  useEffect(() => {
-    if (selectedTag) {
-      fetchPostsByTag(selectedTag);
-    } else {
-      fetchPosts();
-    }
-    updateURL();
-  }, [skip, limit, sortBy, sortOrder, selectedTag]);
-
-  useEffect(() => {
-    const params = new URLSearchParams(location.search);
-    setSkip(parseInt(params.get("skip") || "0"));
-    setLimit(parseInt(params.get("limit") || "10"));
-    setSearchQuery(params.get("search") || "");
-    setSortBy(params.get("sortBy") || "");
-    setSortOrder(params.get("sortOrder") || "asc");
-    setSelectedTag(params.get("tag") || "");
-  }, [location.search]);
 
   // 게시물 테이블 렌더링
 
@@ -305,7 +172,6 @@ const PostsManager = () => {
       <CardContent>
         <div className="flex flex-col gap-4">
           {/* 검색 및 필터 컨트롤 */}
-
           <div className="flex gap-4">
             <div className="flex-1">
               <div className="relative">
